+++ CHANGELOG.md
@@ -22,6 +22,7 @@ at the cost of longer loading times
 #### What's Improved
 - Migrated to better-looking motion transitions
 - App now exposes an (immutable) queue to the MediaSession
+- Improved widget sizing/cover management
 
 #### What's Fixed
 - Fixed default material theme being used before app shows up